refactor module exports
