+++ apps/app/src/server/models/index.js
@@ -22,3 +22,5 @@ export * as PageRedirect from './page-redirect';
 export * as ShareLink from './share-link';
 export * as Tag from './tag';
 export * as UserGroup from './user-group';
+
+export * from './serializers';
