+++ apps/app/src/server/models/serializers/index.ts
@@ -1,0 +1,5 @@
+export * from './bookmark-serializer';
+export * from './page-serializer';
+export * from './revision-serializer';
+export * from './user-group-relation-serializer';
+export * from './user-serializer';
