+++ apps/app/src/server/routes/attachment/index.ts
@@ -1,0 +1,1 @@
+export * from './attachment';
